// app/lib/pdf-templates/types.ts
export interface CreditApplication {
  // Headers & Labels
  mainTitle: string;
  subTitle: string;
  firmNameLabel: string;
  addressLabel: string;
  cityLabel: string;
  stateLabel: string;
  zipLabel: string;
  phoneLabel: string;
  faxLabel: string;
  websiteLabel: string;
  businessInfoTitle: string;
  bankInfoTitle: string;
  tradeReferencesTitle: string;
  agreementTitle: string;
  
  // Data
  firmName: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  phone: string;
  fax: string;
  website: string;
  email: string;
  typeOfBusiness: "Agent" | "NVOCC" | "Forwarder" | "Other";
  apContact: string;
  apTitle: string;
  companyRegState: string;
  federalId: string;
  bankName: string;
  bankAccount: string;
  bankAddress: string;
  bankContact: string;
  tradeReferences: Array<{
    name: string;
    address: string;
    contact: string;
    phone: string;
    email: string;
  }>;
  agreementDate: string;
  signature: string;
  title: string;
}